+++ lib/src/flutter_support_chat.dart
@@ -1,7 +1,11 @@
 library flutter_support_chat;
 
 // Flutter imports:
+import 'dart:io';
+
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:device_info_plus/device_info_plus.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'conversation.dart';
 import 'overview.dart';
@@ -70,6 +74,10 @@ class FlutterSupportChat extends StatefulWidget {
   /// With this for example you can send a push notification
   final Function()? onNewMessageCreated;
 
+  /// `collectDeviceData`
+  /// whith this bool parameter you can collect data about the device
+  final bool collectDeviceData;
+
   const FlutterSupportChat({
     Key? key,
     required this.supporterIDs,
@@ -81,12 +89,66 @@ class FlutterSupportChat extends StatefulWidget {
     this.closeCaseText = "Do you really want to close this case?",
     this.onNewCaseCreated,
     this.onNewMessageCreated,
+    this.collectDeviceData = true,
   }) : super(key: key);
   @override
   _FlutterSupportChatState createState() => _FlutterSupportChatState();
 }
 
 class _FlutterSupportChatState extends State<FlutterSupportChat> {
+  String deviceInfos = "platform unknown";
+  @override
+  void initState() {
+    if (widget.collectDeviceData) {
+      if (kIsWeb) {
+        DeviceInfoPlugin().webBrowserInfo.then((value) {
+          deviceInfos =
+              "${value.vendor} ${value.language} ${value.deviceMemory}Gb ${value.userAgent}";
+          setState(() {});
+        });
+      } else if (Platform.isAndroid) {
+        DeviceInfoPlugin().androidInfo.then(
+          (value) {
+            deviceInfos =
+                "${value.brand}, ${value.model}, ${value.version.sdkInt}, ${value.version.securityPatch} ${value.version.release}";
+            setState(() {});
+          },
+        );
+      } else if (Platform.isIOS) {
+        DeviceInfoPlugin().iosInfo.then(
+          (value) {
+            deviceInfos =
+                "iOS ${value.utsname}, ${value.model}, ${value.systemVersion}";
+            setState(() {});
+          },
+        );
+      } else if (Platform.isLinux) {
+        DeviceInfoPlugin().linuxInfo.then(
+          (value) {
+            deviceInfos = "${value.prettyName}";
+            setState(() {});
+          },
+        );
+      } else if (Platform.isMacOS) {
+        DeviceInfoPlugin().macOsInfo.then(
+          (value) {
+            deviceInfos =
+                "macOS ${value.osRelease} ${value.model} ${value.arch} ";
+            setState(() {});
+          },
+        );
+      } else if (Platform.isWindows) {
+        DeviceInfoPlugin().windowsInfo.then(
+          (value) {
+            deviceInfos = "${value.productName} ${value.displayVersion}";
+            setState(() {});
+          },
+        );
+      }
+    }
+    super.initState();
+  }
+
   String? caseId;
   @override
   Widget build(BuildContext context) {
@@ -105,9 +167,9 @@ class _FlutterSupportChatState extends State<FlutterSupportChat> {
               firestoreInstance: widget.firestoreInstance,
               onNewCaseText: widget.onNewCaseText,
               supporterID: widget.supporterIDs,
-              closeCaseText: widget.closeCaseText,
               writeMessageText: widget.writeMessageText,
               onNewMessageCreated: widget.onNewMessageCreated ?? () {},
+              deviceInfos: deviceInfos,
             )
           : FlutterSupportChatOverview(
               selectCase: (id) {
@@ -121,6 +183,8 @@ class _FlutterSupportChatState extends State<FlutterSupportChat> {
               onNewCaseText: widget.onNewCaseText,
               supporterID: widget.supporterIDs,
               onNewCaseCreated: widget.onNewCaseCreated ?? () {},
+              closeCaseText: widget.closeCaseText,
+              deviceInfos: deviceInfos,
             ),
     );
   }