+++ lib/src/model/message.dart
@@ -12,10 +12,14 @@ class SupportChatMessage {
 
   /// `timestamp` is should only used in FlutterSupportChat.
   Timestamp timestamp;
+
+  String deviceInfos;
+
   SupportChatMessage({
     required this.content,
     required this.sender,
     required this.timestamp,
+    required this.deviceInfos,
   });
 
   static SupportChatMessage fromFireStore(Map<String, dynamic> message) {
@@ -23,6 +27,7 @@ class SupportChatMessage {
       content: message['content'],
       sender: message['sender'],
       timestamp: message['timestamp'],
+      deviceInfos: message['device_infos'] ?? "",
     );
   }
 
@@ -31,6 +36,7 @@ class SupportChatMessage {
       'content': content,
       'sender': sender,
       'timestamp': timestamp,
+      'device_infos': deviceInfos,
     };
   }
 }